Use "spring.mongodb.embedded" properties to enable replica set to try starting cluster with replica set.
Issue is that the `EmbeddedMongoAutoConfiguration` class in `spring-boot-autoconfigure` module that starts the embedded mongo does not enable journaling, which is required for mongo replica set to work. Hence the mongo process fails to start. Following log message (when you try running tests) confirms this:
```
Running wiredTiger without journaling in a replica set is not supported. Make sure you are not using --nojournal and that storage.journal.enabled is not set to 'false'
```