Overridden default MongodConfig to enable journaling.
The test now fails with following error:
```
Command failed with error 10107 (NotWritablePrimary): 'not primary' on server localhost:56192. The full response is {"topologyVersion": {"processId": {"$oid": "61fc46ef1eb4a33802acd85e"}, "counter": 0}, "ok": 0.0, "errmsg": "not primary", "code": 10107, "codeName": "NotWritablePrimary"}
```
This is because the test is not waiting for the replica set to elect a primary and hence is not functional.